Fixes the bug that cached framebuffer may be deleted unexpectedly.
This commit fixes the bug that cached framebuffer of a widget may be deleted unexpectedly in the `Widget::BeginFramebufferUpdates()` method. The bug is fixed by accepting both `width` and `height` parameters in the `float` type so the calculated dimension of the framebuffer would be consistent.

+++ moui/widgets/widget.h
@@ -236,7 +236,7 @@ class Widget {
   // `nullptr` as the `scale_factor` parameter.
   bool BeginFramebufferUpdates(NVGcontext* context,
                                NVGLUframebuffer** framebuffer,
-                               const int width, const int height,
+                               const float width, const float height,
                                float* scale_factor);
 
   // Initializes the environment for rendering in the passed framebuffer based
